pyproject.toml
[tool.poetry]
name = "equipment-failure-prediction"
version = "0.1.0"
description = "API for predicting equipment failure using CatBoost and Streamlit"
authors = ["Gabriel Mota <gabriel.mota@ufv.br>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "api/app" },
    { include = "streamlit_app" },
    { include = "training" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

# Core frameworks
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}

# Data science
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
catboost = "^1.2.5"

# Utilities
joblib = "^1.4.2"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.4"  # Para ler arquivos Excel

# Web app
streamlit = "^1.35.0"

# Typing & validation
pydantic = "^2.7.1"

# Testing (apenas para desenvolvimento)
pytest = {version = "^8.2.2", optional = true}
httpx = {version = "^0.27.0", optional = true}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
flake8 = "^7.0.0"
pytest-mock = "^3.14.0"
unittest-xml-reporting = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-api = "api.app.main:run"